+++ CHANGELOG.md
@@ -38,6 +38,7 @@ Improvements:
 * Locale: you can now import/export locale from/to xml. The installers also use xml's.
 * Locale: export for missing locale.
 * Locale: remove deprecated insertLocale function.
+* Locale: created an incredibly nasty hotfix for some deprecated PHP functionality.
 * Mailmotor: added extra validation (reported on http://forkcms.lighthouseapp.com/projects/61890/tickets/137).
 * Mailmotor: added extra validation for adding address(es).
 * Mailmotor: adding multiple addresses now uses the multipleTextbox-functionality.